MYR-1481: implement create and update interceptor in controller

+++ src/controllers/user-currency.controller.ts
@@ -4,6 +4,7 @@ import {
   del,
   get,
   getModelSchemaRef,
+  HttpErrors,
   param,
   patch,
   post,
@@ -163,6 +164,20 @@ export class UserCurrencyController {
     })
     userCurrency: UserCurrency,
   ): Promise<Count> {
+    const user = await this.userRepository.findById(userCurrency.userId);
+
+    if (user.defaultCurrency === userCurrency.currencyId) {
+      throw new HttpErrors.UnprocessableEntity(
+        'Please changed your default currency, before deleting it',
+      );
+    }
+
+    const {count} = await this.userCurrencyRepository.count({userId: user.id});
+
+    if (count === 1) {
+      throw new HttpErrors.UnprocessableEntity('You cannot delete your only currency');
+    }
+
     return this.userCurrencyRepository.deleteAll({
       userId: userCurrency.userId,
       currencyId: userCurrency.currencyId,
